+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './layout';
 export * from './buildInfo';
 export * from './dialogs';
-export * from './errors';
+export * from './Errors';
 export * from './header';
 export * from './metaData';
 export {default} from './phovea';
